Move myla_globals processsing to globals; pass data as props to App in index.js
