+++ assets/src/index.js
@@ -1,13 +1,18 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter as Router } from 'react-router-dom'
+import { ThemeProvider } from '@material-ui/core/styles'
 import './index.css'
 import App from './containers/App'
+import { user, siteTheme, gaId, loginURL } from './globals'
 // import * as serviceWorker from './serviceWorker'
+
 
 ReactDOM.render(
   <Router basename='/'>
-    <App />
+    <ThemeProvider theme={siteTheme}>
+      <App user={user} gaId={gaId} loginURL={loginURL} />
+    </ThemeProvider>
   </Router>
   , document.getElementById('root')
 )
